rename dynamo service to match consistency of other services
